Show help when the streaming mode is enabled.
Fixes #59
