+++ tests/test_plugin.py
@@ -1,46 +1,4 @@
-import pytest
-
 from tap.tracker import ENABLE_VERSION_13
-
-
-@pytest.fixture
-def sample_test_file(testdir):
-    testdir.makepyfile(
-        """
-        import pytest
-
-        def test_ok():
-            assert True
-
-        def test_not_ok():
-            assert False
-
-        @pytest.mark.parametrize('param', ("foo", "bar"))
-        def test_params(param):
-            assert True
-
-        @pytest.mark.skip(reason='some reason')
-        def test_skipped():
-            assert False
-
-        @pytest.mark.xfail(reason='a reason')
-        def test_broken():
-            assert False
-    """
-    )
-
-
-def test_includes_options(testdir):
-    """All options are present in the help."""
-    result = testdir.runpytest("--help")
-
-    expected_option_flags = [
-        "*--tap-stream*",
-        "*--tap-files*",
-        "*--tap-combined*",
-        "*--tap-outdir=path*",
-    ]
-    result.stdout.fnmatch_lines(expected_option_flags)
 
 
 def test_stream(testdir, sample_test_file):
